Update closest() to accept a css selector Move closest() and once() tests into separate files Add additional once() tests
