+++ lib/closest.js
@@ -1,6 +1,9 @@
 import matches from './matches';
+import findOne from './findOne';
 
-function closest(el, selector) {
+function closest(origin, selector) {
+  const el = (selector instanceof HTMLElement) ? origin : findOne(origin);
+
   let parent = el.parentElement;
   while (parent.parentElement && !matches(parent, selector)) {
     parent = parent.parentElement;
